+++ src/platforms/wordpress/classes/Gantry/WordPress/Assignments/AssignmentsContext.php
@@ -17,6 +17,11 @@ class AssignmentsContext implements AssignmentsInterface
     public $type = 'context';
     public $priority = 1;
 
+    protected $priorities = [
+        'is_front_page'     => 2,
+        'is_home'           => 2,
+    ];
+
     protected $context = [
         'is_404'            => '404 Not Found Page',
         'is_search'         => 'Search Page',
@@ -46,13 +51,13 @@ class AssignmentsContext implements AssignmentsInterface
         $rules = [];
         foreach($this->context as $var => $label) {
             if (isset($wp_query->$var) && $wp_query->$var === true) {
-                $rules[$var] = $this->priority;
+                $rules[$var] = isset($this->priorities[$var]) ? $this->priorities[$var] : $this->priority;
             }
         }
 
         // Workaround for when is_front_page is missing in the $wp_query
         if(is_front_page() === true) {
-            $rules['is_front_page'] = $this->priority;
+            $rules['is_front_page'] = $this->priorities['is_front_page'];
         }
 
         // Allow to filter out rules by 3rd party plugin integration